api/requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.5
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt
djoser==2.1.0
gunicorn==20.1.0
idna==3.2
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
oauthlib==3.1.1
psycopg2-binary==2.9.1
pycparser==2.20
PyJWT
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django
social-auth-core
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.6
django-nose==1.4.7
factory_boy==3.2.0
djangorestframework-jwt
